LookAt矩阵的构建过程

+++ app/src/main/cpp/sample/Transform.cpp
@@ -83,9 +83,9 @@ void Transform::Draw() {
     glm::mat4 trans = glm::mat4(1.0f);
     /*trans = glm::rotate(trans, glm::radians(90.0f), glm::vec3(0.0, 0.0, 1.0));
     trans = glm::scale(trans, glm::vec3(0.5, 0.5, 0.5));*/
-    int time = TimeUtils::currentTimeSeconds();
+    float time = TimeUtils::currentTimeSeconds();
     trans = glm::translate(trans, glm::vec3(0.5f, -0.5f, 0.0f));
-    trans = glm::rotate(trans, static_cast<float>(time), glm::vec3(0.0f, 0.0f, 1.0f));
+    trans = glm::rotate(trans, time, glm::vec3(0.0f, 0.0f, 1.0f));
     setMat4(m_ProgramObj, "transform", trans);
 
     //绑定VAO
